code/chapter_08/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.23.0
openai-agents>=0.0.19
openai>=1.90.0
python-dotenv>=1.0.0
pydantic>=1.10.0